Добавляет отдельный компонент cities-board
