+++ project/src/components/sorting/sorting.tsx
@@ -8,7 +8,7 @@ function Sorting(): JSX.Element  {
           <use xlinkHref="#icon-arrow-select"></use>
         </svg>
       </span>
-      <ul className="places__options places__options--custom places__options--opened">
+      <ul className="places__options places__options--custom">
         <li className="places__option places__option--active" tabIndex={0}>Popular</li>
         <li className="places__option" tabIndex={0}>Price: low to high</li>
         <li className="places__option" tabIndex={0}>Price: high to low</li>
