+++ project/src/store/action.ts
@@ -1,6 +1,7 @@
 import { ActionType, ChangeCityAction } from '../types/action';
+import { City } from '../types/city';
 
-export const changeCity = (city: string): ChangeCityAction => ({
+export const changeCity = (city: City): ChangeCityAction => ({
   type: ActionType.ChangeCity,
   payload: city,
 });
