+++ project/src/types/action.ts
@@ -1,10 +1,12 @@
+import { City } from './city';
+
 export enum ActionType {
   ChangeCity = 'main/changeCity',
 }
 
 export type ChangeCityAction = {
   type: ActionType.ChangeCity;
-  payload: string;
+  payload: City;
 }
